More work on 1v1 matchmaking: - Changed the type of Player.PlayerPosition.Energy to short so that it matches the source, the position packet. - Added the OneVersusOneStats module to track stats for 1v1 matches. Updated the Match1v1 module to fire callbacks for the new module to tie into.

+++ src/Core/Player.cs
@@ -452,7 +452,7 @@ namespace SS.Core
             /// <summary>
             /// current energy
             /// </summary>
-            public int Energy { get; internal set; }
+            public short Energy { get; internal set; }
 
             /// <summary>
             /// time of last position packet
